Fix wrong setpoint read from endpoint in manual setpoint mode
Netatmo, in their infinite wisdom, decided to use two separate properties for the setpoint temperature in the Getthermostatdata endpoint - one for the manual mode, and one for everything else.

The library incorrectly assumes that the setpoint temperature in manual mode should be read from the same property as for other modes. This commit should fix it.